fix: Address Cursor bot feedback - window cleanup and pointer truncation
Fixes two critical issues identified in PR #254 review:

1. HIGH SEVERITY - Pointer Truncation Bug (FIXED):
   - Changed visitprocessthreads signature from int32_t refcon → long refcon
   - On 64-bit systems, long is 8 bytes (holds pointers), int32_t is 4 bytes
   - Existing callers pass (long)ht where ht is pointer, callbacks cast back
   - Truncation to 32-bit would corrupt pointer values

   Files: Common/headers/processinternal.h, Common/source/process.c
   Lines: processinternal.h:272,274, process.c:497

2. MEDIUM SEVERITY - Missing Hidden Window Cleanup (FIXED):
   - Added langclosehiddenwindow() call to langunsettarget() before deletion
   - Original implementation bypassed langcleartarget() which closed hidden windows
   - In GUI builds, this caused resource leak - orphaned hidden editor windows
   - In headless builds, langclosehiddenwindow() is no-op (no impact)

   File: Common/source/langverbs.c
   Implementation pattern:
   - Lookup target value using hashlookup()
   - Call langclosehiddenwindow(val) to close hidden windows (GUI only)
   - Then proceed with hashdelete()
   - Removed unused houterlocaltable variable

Test Results:
- 46/46 target verb integration tests passing
- Build successful with no errors
- Headless behavior unchanged (langclosehiddenwindow is no-op)
- GUI builds will properly close hidden windows when target deleted

+++ Common/source/process.c
@@ -494,7 +494,7 @@ boolean processfindcode (hdltreenode hcode, hdlprocessrecord *hprocess) {
 
 static boolean flvisitingthreads = false; // *** debug
 
-boolean visitprocessthreads (pascal boolean (*visit) (hdlthreadglobals, int32_t), int32_t refcon) {
+boolean visitprocessthreads (pascal boolean (*visit) (hdlthreadglobals, long), long refcon) {
 
 	/*
 	visit all process threads until the visit routine returns true
